% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bootPaq1000}
\alias{bootPaq1000}
\title{bootPaq1000 dataset}
\format{
boostrap model on Paquid data set
}
\usage{
data(bootPaq1000)
}
\description{
Description of the bootPaq1000 dataset.
}
\keyword{datasets}
