% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DYNINLAidmpredY.R
\name{DYNINLAidmpredY}
\alias{DYNINLAidmpredY}
\title{Calculate predictions for time-depend covariates using INLA}
\usage{
DYNINLAidmpredY(
  timeVar,
  times,
  formLong,
  data,
  dataLongi,
  id,
  assoc,
  modelY,
  seed,
  BLUP,
  scale.X
)
}
\description{
Calculate predictions for time-depend covariates using INLA
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
