% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DYNpredY.R
\name{DYNpredY}
\alias{DYNpredY}
\title{Fit an dynamic illness-death model with limited and numerous explanatory variables, including time dependent covariates}
\usage{
DYNpredY(
  formula01,
  formula02,
  formula12,
  data,
  formLong,
  Longitransition,
  threshold,
  dataLongi,
  timeVar,
  id,
  Nsample,
  methodJM = list(functional_forms, n_iter = 3500, n_burnin = 500, n_thin = 1, n_chains =
    3, seed = 1234),
  methodINLA = list(family = "gaussian", basRisk = "rw1", assoc = NULL),
  nproc = 1,
  clustertype = "FORK",
  Ypredmethod = "gauss",
  NtimesPoints = 250,
  envir = parent.frame()
)
}
\arguments{
\item{formula01}{A formula specifying a regression model for the
\code{0 --> 1} transition from the initial state to the transient
state of the illness-death model.  The right hand side of the
formula specifies the covariate terms, and the left hand side must
be an event history object as returned by the function \code{Hist}.}

\item{formula02}{A formula specifying a regression model for the
\code{0 --> 2} transition from the initial state to the absorbing
state. The left hand side must be equal to the left hand side of
\code{formula01}. If missing it is set to \code{formula01}.}

\item{formula12}{A formula specifying a regression model for the
\code{1 --> 2} transition from the transient state to the absorbing
state.  operator is not required. If missing it is set to
\code{formula01}.}

\item{data}{A survival data frame in which to interpret the variables of
\code{formula01}, \code{formula02} and \code{formula12}.}

\item{dataLongi}{A longitudinal data frame for the time dependent covariates
This argument is only used for models with M-splines.}

\item{nproc}{number of cluster}

\item{clustertype}{in which cluster to work}

\item{envir}{working environment}

\item{na.action}{how NAs are treated. The default is first, any
na.action attribute of data, second a na.action setting of options,
and third 'na.fail' if that is unset. The 'factory-fresh' default
is na.omit. Another possible value is NULL.}
}
\description{
Fit an illness-death model with a limited number of explanatory variables
 using semi-parametric approach with a linear combination of M-splines or 
 Weibull distribution on the transition intensities. 
Fit an illness-death model with a large number of explanatory variables 
using regularization approach of the log-likelihood, i.e. lasso, elastic-net,
 MCP or SCAD penalty.
Left-truncated, right-censored, and interval-censored data are allowed.
State 0 corresponds to the initial state, state 1 to the transient one,
state 2 to the absorbant one. The allowed transitions are: 0 --> 1, 0 --> 2
and 1 --> 2.

The estimated parameters are obtained using the robust Marquardt algorithm
(Marquardt, 1963) or a proximal gradient algorithm.
}
\references{
D. Marquardt (1963). An algorithm for least-squares estimation
of nonlinear parameters.  \emph{SIAM Journal of Applied Mathematics},
431-441.
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
\keyword{illness-death}
