% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictYtimespoints.R
\name{gauss_kronrod_points}
\alias{gauss_kronrod_points}
\title{Define the time points at which Y need to be estimated}
\usage{
gauss_kronrod_points(
  lower.intdouble,
  upper.intdouble,
  end.time,
  truncation,
  entry.time
)
}
\arguments{
\item{lower.intdouble}{lower bound of integral for double integral estimation}

\item{upper.intdouble}{upper bound of integral for double integral estimation}

\item{end.time}{upper bound of integral from 0 to end of follow-up (death or censoring)}

\item{truncation}{TRUE if left-truncation, otherwise FALSE}

\item{entry.time}{time of entry in the study (values 0 if no left truncation, otherwise >0)}
}
\description{
Define the time points at which Y need to be estimated
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
