% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DYNidmlLikelihoodpena.R
\name{gaussDYNidmlLikelihoodpena}
\alias{gaussDYNidmlLikelihoodpena}
\title{Log likelihood with M-spline baseline risk}
\usage{
gaussDYNidmlLikelihoodpena(
  b,
  npm,
  npar,
  bfix,
  fix,
  zi01,
  zi02,
  zi12,
  ctime,
  no,
  nz01,
  nz02,
  nz12,
  ve01,
  ve02,
  ve12,
  dimnva01,
  dimnva02,
  dimnva12,
  nva01,
  nva02,
  nva12,
  t0,
  t1,
  t2,
  t3,
  troncature,
  y01,
  y02,
  y12,
  p01,
  p02,
  p12,
  dimp01,
  dimp02,
  dimp12,
  Ntime,
  time,
  lambda,
  alpha,
  penalty.factor,
  penalty
)
}
\arguments{
\item{b}{parameters not fixed}

\item{npm}{number of parameters not fixed}

\item{npar}{number of parameters}

\item{bfix}{parameters fixed}

\item{fix}{indicators of fixed and unfixed parameters}

\item{ctime}{classification of subject according to their observations}

\item{no}{number of subjects}

\item{ve01}{variables for transition 0 -->1}

\item{ve02}{variables for transition 0 -->2}

\item{ve12}{variables for transition 1 -->2}

\item{dimnva01}{number of variables for transition 0 -->1}

\item{dimnva02}{number of variables for transition 0 -->2}

\item{dimnva12}{number of variables for transition 1 -->2}

\item{nva01}{number of variables for transition 0 -->1}

\item{nva02}{number of variables for transition 0 -->2}

\item{nva12}{number of variables for transition 1 -->2}

\item{t0}{time entry}

\item{t1}{time L}

\item{t2}{time R}

\item{t3}{time of event/out}

\item{troncature}{indicator if troncature or not}

\item{lambda}{matrix of penalty lambda with column one for transiton from 
0 --> 1, column two from 0 --> 2 and column 3 for 1 --> 2}

\item{alpha}{alpha on all transitions}

\item{penalty.factor}{which variable should be penalised}

\item{penalty}{which penalty to consider}

\item{gausspoint}{number of points in gauss quadrature}
}
\description{
Log likelihood with M-spline baseline risk
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
