% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIDeM-package.R
\docType{package}
\name{HIDeM}
\alias{HIDeM-package}
\alias{HIDeM}
\title{HIDeM: High-dimensional Illness-death Models}
\description{
This package provides tools for fitting regularized illness-death models 
with interval-censored data.
}
\details{
\strong{Main Functions:}
\itemize{
  \item{\code{idm()}: Fit the illness-death model with variable selection.}
  \item{\code{simulateIDM()}: Simulate data from an illness-death model with      #' interval censored event times and covariates.}
  \item{\code{intensity()}: estimate transition intensities}
  \item{\code{predict()}: predict life-time expectancy, survival and transition intensities.}
  \item{\code{plot}: Plot estimated baseline transition intensities.}
  \item{\code{summary()}: Summarize model results.}
}
}
\keyword{package}
