% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLAidm.R
\name{INLAidm}
\alias{INLAidm}
\title{Calculate predictions for time-depend covariates using INLA}
\usage{
INLAidm(
  timeVar,
  family,
  basRisk,
  assoc,
  truncated,
  formLong,
  formSurv,
  dataSurv,
  dataLongi,
  id,
  NsampleHY,
  NsampleFE,
  NsampleRE,
  nproc,
  t0,
  t1,
  t2,
  t3,
  idm,
  idd,
  clustertype,
  Ypredmethod,
  NtimesPoints
)
}
\description{
Calculate predictions for time-depend covariates using INLA
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
