% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLAidmpredY.R
\name{INLAidmpredY}
\alias{INLAidmpredY}
\title{Calculate predictions for time-depend covariates using INLA}
\usage{
INLAidmpredY(
  timeVar,
  truncated,
  formLong,
  dataSurv,
  dataLongi,
  id,
  Nsample,
  t0,
  t1,
  t2,
  t3,
  ctime,
  modelY,
  seed,
  BLUP
)
}
\description{
Calculate predictions for time-depend covariates using INLA
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
