% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMidm.R
\name{JMidm}
\alias{JMidm}
\title{Calculate predictions for time-depend covariates using INLA}
\usage{
JMidm(
  timeVar,
  functional_forms,
  truncated,
  formLong,
  formSurv,
  dataSurv,
  dataLongi,
  id,
  n_iter,
  n_burnin,
  n_thin,
  n_chain,
  nproc,
  t0,
  t1,
  t2,
  t3,
  idm,
  idd,
  clustertype,
  lightmode
)
}
\description{
Calculate predictions for time-depend covariates using INLA
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
