% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m01example}
\alias{m01example}
\title{m01example dataset}
\format{
model on transition specific 0 to 1 using data set simulated
}
\usage{
data(m01example)
}
\description{
Description of the m01example dataset.
}
\keyword{datasets}
