% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m01Paq1000}
\alias{m01Paq1000}
\title{m01Paq1000 dataset}
\format{
model on transition specific 0 to 1 using Paquid data set
}
\usage{
data(m01Paq1000)
}
\description{
Description of the m01Paq1000 dataset.
}
\keyword{datasets}
