% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m02example}
\alias{m02example}
\title{m02example dataset}
\format{
model on transition specific 0 to 2 using data set simulated
}
\usage{
data(m02example)
}
\description{
Description of the m02example dataset.
}
\keyword{datasets}
