% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m02Paq1000}
\alias{m02Paq1000}
\title{m02Paq1000 dataset}
\format{
model on transition specific 0 to 2 using Paquid data set
}
\usage{
data(m02Paq1000)
}
\description{
Description of the m02Paq1000 dataset.
}
\keyword{datasets}
