% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m12example}
\alias{m12example}
\title{m12example dataset}
\format{
model on transition specific 1 to 2 using data set simulated
}
\usage{
data(m12example)
}
\description{
Description of the m12example dataset.
}
\keyword{datasets}
