% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m12Paq1000}
\alias{m12Paq1000}
\title{m12Paq1000 dataset}
\format{
model on transition specific 1 to 2 using Paquid data set
}
\usage{
data(m12Paq1000)
}
\description{
Description of the m12Paq1000 dataset.
}
\keyword{datasets}
