% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modelexample}
\alias{modelexample}
\title{modelexample dataset}
\format{
Final model on simulated data set
}
\usage{
data(modelexample)
}
\description{
Description of the modelexample dataset.
}
\keyword{datasets}
