% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modelPaq1000}
\alias{modelPaq1000}
\title{modelPaq1000 dataset}
\format{
Final model on Paquid data set
}
\usage{
data(modelPaq1000)
}
\description{
Description of the modelPaq1000 dataset.
}
\keyword{datasets}
