% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvoidexample}
\alias{mvoidexample}
\title{mvoidexample dataset}
\format{
void model on simulated data set
}
\usage{
data(mvoidexample)
}
\description{
Description of the mvoidexample dataset.
}
\keyword{datasets}
