% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvoidPaq1000}
\alias{mvoidPaq1000}
\title{mvoidPaq1000 dataset}
\format{
void model on Paquid data set
}
\usage{
data(mvoidPaq1000)
}
\description{
Description of the mvoidPaq1000 dataset.
}
\keyword{datasets}
