% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictYtimespoints.R
\name{predict_for_id}
\alias{predict_for_id}
\title{Define the time points at which Y need to be estimated if we have truncation}
\usage{
predict_for_id(id, truncation, seen.ill, seen.exit)
}
\arguments{
\item{a}{lower bound of integral}

\item{b}{upper bound of integral}
}
\description{
Define the time points at which Y need to be estimated if we have truncation
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
