% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remodelexample}
\alias{remodelexample}
\title{remodelexample dataset}
\format{
Re estimated final model on simulated data set
}
\usage{
data(remodelexample)
}
\description{
Description of the remodelexample dataset.
}
\keyword{datasets}
