% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remodelPaq1000}
\alias{remodelPaq1000}
\title{remodelPaq1000 dataset}
\format{
Re estimated final model on Paquid data set
}
\usage{
data(remodelPaq1000)
}
\description{
Description of the remodelPaq1000 dataset.
}
\keyword{datasets}
