% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMidmpredY.R
\name{JMidmpredY}
\alias{JMidmpredY}
\title{Calculate predictions for time-depend covariates using INLA}
\usage{
JMidmpredY(
  timeVar,
  truncated,
  formLong,
  dataSurv,
  dataLongi,
  id,
  t0,
  t1,
  t2,
  t3,
  ctime,
  assoc,
  modelY,
  seed,
  BLUP,
  scale.X
)
}
\description{
Calculate predictions for time-depend covariates using INLA
}
\author{
R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>
}
